refactor Dropper
